# server/swagger/mastermind-v3.json
{
    "swagger": "2.0",
    "info": {
        "title": "MasterMind Middleware API",
        "description": "The data storage middleware for MasterMind",
        "version": "3.0.0"
    },
    "host": "localhost:3000",
    "schemes": [
        "http"
    ],
    "basePath": "/v3",
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/vacations": {
            "x-handler": "Vacations",
            "get": {
                "summary": "Retrieve Vacations",
                "description": "The Vacations endpoint returns vacations.\n",
                "operationId": "getVacations",
                "tags": [
                    "Vacations"
                ],
                "parameters": [
                    {
                        "name": "people",
                        "in": "query",
                        "description": "Filter vacations by only those belonging to one of these people (specified by the ID of the people)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "statuses",
                        "in": "query",
                        "description": "Filter vacations to only those in one of the given statuses",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "enum": [
                                "Approved",
                                "Pending",
                                "Cancelled",
                                "Denied"
                            ]
                        }
                    },
                    {
                        "name": "manager",
                        "in": "query",
                        "description": "Filter vacations to only those which have this person as the manager (specified by the ID of the manager Person)",
                        "type": "string"
                    },
                    {
                        "name": "startDate",
                        "in": "query",
                        "description": "Filter vacations to those starting on or after the given date.",
                        "type": "string",
                        "format": "date-time"
                    },
                    {
                        "name": "endDate",
                        "in": "query",
                        "description": "Filter vacations to those ending before the given date.",
                        "type": "string",
                        "format": "date-time"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An array of vacations",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Vacation"
                            }
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to view Vacations",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "summary": "Create Vacation",
                "description": "Create a new vacation.\nAn ID in the POST-ed vacation will be ignored.\n",
                "operationId": "createSingleVacation",
                "tags": [
                    "Vacations"
                ],
                "parameters": [
                    {
                        "name": "vacation",
                        "in": "body",
                        "description": "Data representing the vacation to create",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Vacation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The newly created vacation",
                        "schema": {
                            "$ref": "#/definitions/Vacation"
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to create new Vacations",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/vacations/{id}": {
            "x-handler": "Vacations",
            "get": {
                "summary": "Retrieve Vacation",
                "description": "The Vacation endpoint returns a single vacation.\n",
                "operationId": "getSingleVacation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vacation to retrieve.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Vacations"
                ],
                "responses": {
                    "200": {
                        "description": "The single vacation specified by the ID",
                        "schema": {
                            "$ref": "#/definitions/Vacation"
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to view Vacations",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "A vacation with the specified ID could not be found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update Vacation",
                "description": "Updates an existing vacation.\n",
                "operationId": "updateSingleVacation",
                "tags": [
                    "Vacations"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vacation to update.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "vacation",
                        "in": "body",
                        "description": "New data representing the vacation",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Vacation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The newly updated vacation",
                        "schema": {
                            "$ref": "#/definitions/Vacation"
                        }
                    },
                    "403": {
                        "description": "The current user is not allowed to update Vacations",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "A vacation with the specified ID could not be found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete Vacation",
                "description": "Deletes a single vacation.\n",
                "operationId": "deleteSingleVacation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the vacation to delete.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Vacations"
                ],
                "responses": {
                    "200": {
                        "description": "The single vacation specified by the ID was deleted"
                    },
                    "403": {
                        "description": "The current user is not allowed to delete Vacations",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "A vacation with the specified ID could not be found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Vacation": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "endDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "person": {
                    "type": "string",
                    "description": "ID of Person"
                },
                "manager": {
                    "type": "string",
                    "description": "ID of Person"
                },
                "description": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "Approved",
                        "Pending",
                        "Cancelled",
                        "Denied"
                    ]
                },
                "days": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            },
            "required": [
                "person",
                "manager",
                "type",
                "status",
                "startDate",
                "endDate"
            ]
        },
        "Error": {
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "error": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "oauth2": {
            "type": "oauth2",
            "authorizationUrl": "http://localhost:3000/auth/google",
            "flow": "implicit",
            "scopes": {
                "https://www.googleapis.com/auth/plus.login": "authenticated via Google Plus"
            }
        }
    },
    "security": [
        {
            "oauth2": [
                "https://www.googleapis.com/auth/plus.login"
            ]
        }
    ]
}